feat: add hover effect to social media icons
